+++ CHANGELOG.md
@@ -16,6 +16,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 * Removed `CanFilter` and `CanSort` properties. These can still be set in your code, but will no longer pollute your entities.
 * **TODO update querykit to non-prerelease**
 
+### Fixed
+
+* Sql server setups can handle dateonly
+
 ## [0.21.1] - 07/20/2023
 
 ### Fixed